Consolidate Python-level data
