+++ tools/CMakeLists.txt
@@ -2,7 +2,6 @@ file(
   COPY
     go_pcm.py
     plot_cavity.py
-    setup.py
   DESTINATION
     ${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}
   )
@@ -10,7 +9,6 @@ install(
   FILES
     ${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}/go_pcm.py
     ${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}/plot_cavity.py
-    ${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}/setup.py
   DESTINATION
     ${CMAKE_INSTALL_BINDIR}
   )
@@ -23,7 +21,7 @@ endif()
 configure_file(init.py ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/__init__.py COPYONLY)
 file(
   COPY
-    codata.py
+    pcmdata.py
     pcmparser.py
     pyparsing.py
   DESTINATION
@@ -32,7 +30,7 @@ file(
 install(
   FILES
     ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/__init__.py
-    ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/codata.py
+    ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/pcmdata.py
     ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/getkw.py
     ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/pcmparser.py
     ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/pyparsing.py
